feat: 기기별대응 위해 추가
