+++ LovelyCrane/Ext+/Color+.swift
@@ -1,0 +1,28 @@
+//
+//  Color+Ext.swift
+//  LoveCrane
+//
+//  Created by 최진용 on 2023/07/17.
+//
+
+import SwiftUI
+
+
+extension Color {
+    static let fontGray = Color(UIColor(hexCode: "8F8F8F"))
+    static let white = Color(UIColor(hexCode: "#FFFFFF"))
+    static let backGround = Color(UIColor(hexCode: "#252526"))
+    
+    static let buttonPink = Color(UIColor(hexCode: "FF76B4"))
+    static let fontYellow = Color(UIColor(hexCode: "FFD60A"))
+    
+    static let textFieldGray = Color(UIColor(hexCode: "4A4848"))
+    static let buttonGray = Color(UIColor(hexCode: "3A3838"))
+    static let darkFontGray = Color(UIColor(hexCode: "665F5F"))
+    static let fontBrown = Color(UIColor(hexCode: "524343"))
+
+    static let notReadCell = Color(UIColor(hexCode: "#444444"))
+    static let readCell = Color(UIColor(hexCode: "#303030"))
+    static let blue = Color(.blue)
+
+}
