--- LovelyCrane/Ext+/Color.swift
@@ -1,28 +1,0 @@
-//
-//  Color+Ext.swift
-//  LoveCrane
-//
-//  Created by 최진용 on 2023/07/17.
-//
-
-import SwiftUI
-
-
-extension Color {
-    static let fontGray = Color(UIColor(hexCode: "8F8F8F"))
-    static let white = UIColor(hexCode: "#FFFFFF")
-    static let backGround = UIColor(hexCode: "#252526")
-    
-    static let buttonPink = UIColor(hexCode: "FF76B4")
-    static let fontYellow = UIColor(hexCode: "FFD60A")
-    
-    static let textFieldGray = UIColor(hexCode: "4A4848")
-    static let buttonGray = UIColor(hexCode: "3A3838")
-    static let darkFontGray = UIColor(hexCode: "665F5F")
-    static let fontBrown = UIColor(hexCode: "524343")
-
-    static let notReadCell = Color(UIColor(hexCode: "#444444"))
-    static let readCell = Color(UIColor(hexCode: "#303030"))
-    static let blue = Color(.blue)
-
-}
