+++ LovelyCrane/Ext+/UIScreen+.swift
@@ -5,7 +5,7 @@
 //  Created by 235 on 2023/07/25.
 //
 
-import Foundation
+import UIKit
 extension UIScreen {
     static let screenWidth = UIScreen.main.bounds.size.width
     static let screenHeight = UIScreen.main.bounds.size.height
